Renammed Generator.java to MockLogger.java Added a method to stream mock log lines to a file, and made called it in the main()
